+++ config/config.go
@@ -1,0 +1,5 @@
+package config
+
+const (
+	APIKeyHeader = "X-API-Key"
+)
